broke out analyze-features.ts to its own directory
