+++ packages/scripts/src/analyze-feature/index.ts
@@ -1,0 +1,1 @@
+export { loadFeaturesFromPackages } from './analyze-feature'
